Move `updateLogs` & `createLogs` listeners to `App`
* Move `updateLogs` & `createLogs` listeners to `App` (src/renderer/App.tsx, src/renderer/components/templates/ProxyLog/ProxyLog.tsx);
* Update `Layout`: display **logsActor** state (src/renderer/components/layout/Layout.tsx).
