+++ src/renderer/App.tsx
@@ -2,6 +2,7 @@ import { useEffect } from "react";
 import { uiActor, proxiesActor, logsActor } from "@/xstate";
 import Layout from "@/renderer/components/layout";
 import debug from "debug";
+import { ILogsRecord } from "@/types";
 debug.enable("*");
 
 const logger = debug("renderer:App");
@@ -18,7 +19,37 @@ const App = () => {
         logger(proxy);
       });
     });
-  }, []); // eslint-disable-line react-hooks/exhaustive-deps
+  }, []); // eslint-disable-line
+
+  useEffect(() => {
+    const removeEventListener = window.electronAPI.updateLogs(
+      (value: ILogsRecord) => {
+        logger(value);
+        // if (logId !== value.proxyId) return;
+
+        logsActor.send({ type: "update", updatedLog: value });
+      }
+    );
+
+    return () => {
+      removeEventListener();
+    };
+  }, []); // eslint-disable-line
+
+  useEffect(() => {
+    const removeEventListener = window.electronAPI.createLogs(
+      (value: ILogsRecord) => {
+        logger(value);
+        // if (logId !== value.proxyId) return;
+
+        logsActor.send({ type: "create", newLog: value });
+      }
+    );
+
+    return () => {
+      removeEventListener();
+    };
+  }, []); // eslint-disable-line
 
   useEffect(() => {
     const handleEsc = (event: KeyboardEvent) => {
