+++ src/renderer/components/templates/ProxyLog/ProxyLog.tsx
@@ -3,45 +3,15 @@ import { uiActor, logsActor } from "@/xstate";
 import { XCircle, Trash2 } from "lucide-react";
 import { useSelector } from "@xstate/react";
 import debug from "debug";
-import { useEffect } from "react";
-import { ILogsRecord } from "@/types";
 import { LogsRecord } from "@/renderer/components/templates";
 const logger = debug("Renderer:ProxyLog");
 
 const ProxyLog = () => {
   const logId = useSelector(uiActor, (state) => state.context.logId);
 
+  logger("logId:", logId);
+
   const logs = useSelector(logsActor, (state) => state.context.logs);
-
-  useEffect(() => {
-    const removeEventListener = window.electronAPI.updateLogs(
-      (value: ILogsRecord) => {
-        logger(value);
-        if (logId !== value.proxyId) return;
-
-        logsActor.send({ type: "update", updatedLog: value });
-      }
-    );
-
-    return () => {
-      removeEventListener();
-    };
-  }, []); // eslint-disable-line
-
-  useEffect(() => {
-    const removeEventListener = window.electronAPI.createLogs(
-      (value: ILogsRecord) => {
-        logger(value);
-        if (logId !== value.proxyId) return;
-
-        logsActor.send({ type: "create", newLog: value });
-      }
-    );
-
-    return () => {
-      removeEventListener();
-    };
-  }, []); // eslint-disable-line
 
   return (
     <div className="flex flex-col items-start gap-4 p-2 border rounded-md w-full">
